move formattedError to runtime-common
