+++ packages/runtime-common/error.ts
@@ -1,4 +1,5 @@
 import { getReasonPhrase } from 'http-status-codes';
+import status from 'statuses';
 import { createResponse } from './create-response';
 import { RequestContext } from './realm';
 import type { SearchResultError } from './realm-index-query-engine';
@@ -42,6 +43,84 @@ export interface CardErrorJSONAPI {
 
 export interface CardErrorsJSONAPI {
   errors: CardErrorJSONAPI[];
+}
+
+export function formattedError(
+  url: string | undefined,
+  error: any,
+  err?: CardError | Partial<CardErrorJSONAPI>,
+): CardErrorsJSONAPI {
+  if (isCardError(err)) {
+    let cardError;
+    let meta: CardErrorJSONAPI['meta'] | undefined;
+    let message: string | undefined;
+    let additionalError = err.additionalErrors?.[0];
+
+    if (additionalError && 'errorDetail' in additionalError) {
+      cardError = additionalError.errorDetail;
+      let { lastKnownGoodHtml, scopedCssUrls, cardTitle } = additionalError;
+      meta = {
+        lastKnownGoodHtml,
+        scopedCssUrls,
+        stack: cardError.stack ?? err.stack ?? error.stack ?? null,
+        cardTitle,
+      };
+    } else if (isCardError(additionalError)) {
+      cardError = additionalError;
+    } else {
+      message = additionalError?.message;
+      cardError = err;
+    }
+
+    return {
+      errors: [
+        {
+          id: url,
+          message: message ?? cardError.message,
+          status: cardError.status,
+          title:
+            cardError.title ??
+            status.message[cardError.status] ??
+            cardError.message,
+          realm: error.responseHeaders?.get('X-Boxel-Realm-Url'),
+          meta: meta ?? {
+            lastKnownGoodHtml: null,
+            scopedCssUrls: [],
+            stack: cardError.stack ?? err.stack ?? error.stack ?? null,
+            cardTitle: null,
+          },
+          additionalErrors: cardError.additionalErrors,
+        },
+      ],
+    };
+  }
+
+  let errorStatus = err?.status ?? error.status;
+  let errorMessage =
+    err?.message ??
+    (errorStatus
+      ? `Received HTTP ${errorStatus} from server ${
+          error.responseText ?? ''
+        }`.trim()
+      : `${error.message}: ${error.stack}`);
+
+  return {
+    errors: [
+      {
+        id: url,
+        status: errorStatus ?? '500',
+        title: err?.title ?? status.message[errorStatus] ?? errorMessage,
+        message: errorMessage,
+        realm: err?.realm ?? error.responseHeaders?.get('X-Boxel-Realm-Url'),
+        meta: err?.meta ?? {
+          lastKnownGoodHtml: null,
+          scopedCssUrls: [],
+          stack: error.stack ?? null,
+          cardTitle: null,
+        },
+      },
+    ],
+  };
 }
 
 export class CardError extends Error implements SerializedError {
