+++ packages/runtime-common/package.json
@@ -61,9 +61,11 @@
     "@types/dompurify": "^3.0.2",
     "@types/js-string-escape": "^1.0.1",
     "@types/qunit": "^2.11.3",
+    "@types/statuses": "^2.0.5",
     "concurrently": "^8.2.2",
     "dompurify": "^3.0.3",
-    "ember-cli-htmlbars": "^6.3.0"
+    "ember-cli-htmlbars": "^6.3.0",
+    "statuses": "^2.0.1"
   },
   "peerDependencies": {
     "@babel/core": "^7.26.10",
